Issue 460: fix for warning C4273 on windows (#461)
* remove DLL_EXPORT for from_buffer

* fix more windows built warnings

* warning C4267: 'argument': conversion from 'size_t' to 'int'

* fix warning C4996

+++ src/h5cpp/property/dataset_creation.cpp
@@ -104,7 +104,9 @@ DatasetLayout DatasetCreationList::layout() const {
 }
 
 void DatasetCreationList::chunk(const Dimensions &chunk_dims) const {
-  if (H5Pset_chunk(static_cast<hid_t>(*this), chunk_dims.size(), chunk_dims.data()) < 0) {
+  if (H5Pset_chunk(static_cast<hid_t>(*this),
+		   static_cast<int>(chunk_dims.size()),
+		   chunk_dims.data()) < 0) {
     error::Singleton::instance().throw_with_stack("Failure setting chunk dimensions!");
   }
 }
